requirements.txt
# JobSensei Resume Matching System Requirements

# PDF Processing
pymupdf
PyPDF2
python-docx

# NLP and Text Processing
spacy
textblob

# Semantic Similarity
sentence-transformers
torch

# Data Processing
pandas
numpy

# AI/ML API
groq

# Data Validation and Settings
pydantic
pydantic[email]

# Environment Variables
python-dotenv

# Retry Logic and Resilience
tenacity

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Authentication (built-in modules)
# No additional packages needed - uses hashlib, json, os, sys, dataclasses, datetime

# Computer Vision (for Level 1 CV integration)
opencv-python
mediapipe

# Speech Processing (for Level 1 STT integration)
speechrecognition
pyaudio
